Use different hash function
